backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
anthropic==0.25.0
librosa==0.10.1
websockets==12.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.6.0
pytest==7.4.3
httpx==0.26.0
pytest-asyncio==0.23.0
scipy>=1.10.0
mido>=1.3.0
